Recruitment, mortality, and growth.

These functions are inherited from the the CTFS-R package. Compared to the original functions, these ones have a similar interface but use more conservative defaults and allow suppressing messages. These functions also feature formal tests, bug fixes, additional assertions, and improved messages. (These functions are the internal implementation of their in-development analogs -- hence the suffix `_impl`.)

recruitment_impl(census1, census2, mindbh = NULL, alivecode = NULL,
  split1 = NULL, split2 = NULL, quiet = FALSE)

mortality_impl(census1, census2, alivecode = NULL, split1 = NULL,
  split2 = NULL, quiet = FALSE)

growth_impl(census1, census2, rounddown = FALSE, method = "I",
  stdev = FALSE, dbhunit = "mm", mindbh = NULL, growthcol = "dbh",
  err.limit = 1000, maxgrow = 1000, split1 = NULL, split2 = NULL,
  quiet = FALSE)

Arguments

census1, census2
Two census tables, each being a dataframe and in particular, a ForestGEO tree table. You may use a stem table, but you more commonly should use a tree table because demography analyses make more sense at the scale of a tree than at the scale of stems.

mindbh
The minimum diameter above which the counts are done. Trees smaller than `mindbh` are excluded. If `NULL`, all living trees are included.

alivecode
Character, codes of the variable `status` that indicate the tree is alive. The default 'A' is the standard CTFS designation for living trees or stems.

split1, split2
Optional vectors (column of any one the census dataframe) to aggregate results by. Defaults to aggregating across the entire census datasets.

quiet
Use `TRUE` to suppress messages.

rounddown
If `TRUE``, all `dbh < 55` are rounded down to the nearest multiple of 5.

method
Use "I" to calculate annual dbh increment as `(dbh2 - dbh1)/time`, or "E" to calculate the relative growth rate as `(log(dbh2) - log(dbh1)) / time`.

stdev
Logical. Default (`FALSE`) returns confidence limits, otherwise returns the SD in growth rate per group.

dbhunit
'cm' or 'mm'.

growthcol
defines how growth is measured, either 'dbh' or 'agb' (above ground biomass).

err.limit
A number. Numbers such as 10000 are high and will return all measures.

maxgrow
A number. Numbers such as 10000 are high and will return all measures.

Value

Metrics of recruitment: Similar to metrics of mortality.

Metrics of mortality: * `N`: the number of individuals alive in the census 1 per category selected. * `D`: the number of individuals no longer alive in census 2. * `rate`: the mean annualized mortality rate constant per category selected, calculated as (log(N)-log(S))/time. * `upper`: upper confidence limit of mean rate. * `lower`: lower confidence limit of mean rate. * `time`: mean time interval in years. * `date1`: mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer). * `date2`: mean date in census 2. * `dbhmean`: mean dbh in census 1 of individuals included.

Metrics of growth: * `rate`, the mean annualized growth rate per category selected, either dbh increment, or relative growth * `N`, the number of individuals included in the mean (not counting any excluded) * `clim` (or sd with `stdev = TRUE`), width of confidence interval; add this number to the mean rate to get upper confidence limit, substract to get lower * `dbhmean`, mean dbh in census 1 of individuals included * `time`, mean time interval in years * `date1`, mean date included individuals were measured in census 1, as julian object (R displays as date, but treats as integer) * `date2`, mean date in census 2.

Details

Survivors are all individuals alive in both censuses, with `status == A` in the first census, and larger than the minimum dbh in the first census. The total population in the second census includes all those alive plus any other survivors. Individuals whose status is NA in either census are deleted from all calculations.

Examples

census1 <- fgeo.x::tree5
census2 <- fgeo.x::tree6

recruitment_impl(census1, census2)

#> Detected dbh ranges:
#>   * `census1` = 14-661.
#>   * `census2` = 14.5-676.
#> Using dbh `mindbh = 0` and above.

#> $N2
#> [1] 25
#> 
#> $R
#> [1] 0
#> 
#> $rate
#> [1] 0
#> 
#> $lower
#> [1] 0
#> 
#> $upper
#> [1] 0.03132581
#> 
#> $time
#> [1] 4.529172
#> 
#> $date1
#> [1] 18936.1
#> 
#> $date2
#> [1] 20589.76
#> 

# Demography by any number of grouping variables via `interaction(...)`
sp_quadrat <- interaction(census1$sp, census1$quadrat)

recruitment <- recruitment_impl(
  census1, census2,
  split1 = sp_quadrat,
  quiet = TRUE
)
lapply(recruitment, head)

#> $N2
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>           0           1           1           1           1           1 
#> 
#> $R
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>           0           0           0           0           0           0 
#> 
#> $rate
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA           0           0           0           0           0 
#> 
#> $lower
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA           0           0           0           0           0 
#> 
#> $upper
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA   0.4112830   0.4133016   0.4043707   0.4095329   0.4120377 
#> 
#> $time
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA    4.484600    4.462697    4.561259    4.503765    4.476386 
#> 
#> $date1
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>       18792       18918       18934       18961       19058       19011 
#> 
#> $date2
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA       20556       20564       20627       20703       20646 
#> 

mortality <- mortality_impl(
  census1, census2, split1 = sp_quadrat, quiet = TRUE
)
lapply(mortality, head)

#> $N
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>           1           1           1           1           1           1 
#> 
#> $D
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>           1           0           0           0           0           0 
#> 
#> $rate
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>         Inf           0           0           0           0           0 
#> 
#> $lower
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>  0.03735197  0.00000000  0.00000000  0.00000000  0.00000000  0.00000000 
#> 
#> $upper
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>         Inf   0.4112830   0.4133016   0.4043707   0.4095329   0.4120377 
#> 
#> $time
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>    4.607803    4.484600    4.462697    4.561259    4.503765    4.476386 
#> 
#> $date1
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>       18792       18918       18934       18961       19058       19011 
#> 
#> $date2
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>       20475       20556       20564       20627       20703       20646 
#> 
#> $dbhmean
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>        29.4       106.0        30.5        73.0       661.0        14.0 
#> 

growth <- growth_impl(census1, census2, split1 = sp_quadrat, quiet = TRUE)
lapply(growth, head)

#> $rate
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA   1.5608974   0.2240798   0.2192377   3.3305471   0.4467890 
#> 
#> $N
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>           0           1           1           1           1           1 
#> 
#> $clim
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA          NA          NA          NA          NA          NA 
#> 
#> $dbhmean
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA       106.0        30.5        73.0       661.0        14.0 
#> 
#> $time
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA    4.484600    4.462697    4.561259    4.503765    4.476386 
#> 
#> $date1
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA       18918       18934       18961       19058       19011 
#> 
#> $date2
#>  MIRGAR.102 PREMON.1110  INGLAU.115  IXOFER.117 BUCTET.1223 PSYBRA.1318 
#>          NA       20556       20564       20627       20703       20646 
#> 

# NOT RUN {
# Convert to convenient dataframes -------------------------------------
are_installed <- requireNamespace("fgeo.tool") && requireNamespace("tidyr")
if (are_installed) {
  library(fgeo.tool)
  library(tidyr)

  to_df(
    recruitment_impl(census1, census2, quiet = TRUE)
  )

  to_df(mortality)

  separate(
    to_df(growth),
    groups, into = c("sp", "quadrat")
  )
}
# }