fgeo.analyze: Analyze ForestGEO Data

To help you access, transform, analyze, and visualize ForestGEO data, we developed a collection of R packages (<https://forestgeo.github.io/fgeo/>). This package, in particular, helps you to implement analyses of plot species distributions, topography, demography, and biomass. It also includes a torus translation test to determine habitat associations of tree species as described by Zuleta et al. (2018) <doi:10.1007/s11104-018-3878-0>. To learn more about ForestGEO visit <https://forestgeo.si.edu/>.

See also

Author

Maintainer: Mauro Lepore maurolepore@gmail.com (ORCID) [contractor]

Authors:

Gabriel Arellano gabriel.arellano.torres@gmail.com [reviewer]

Richard Condit richardcondit@gmail.com

Matteo Detto teoparaglider@gmail.com

Kyle Harms kharms@lsu.edu

Suzanne Lao laoz@si.edu [reviewer]

Sabrina Russo srusso2@unl.edu [reviewer]

Daniel Zuleta dfzuleta@gmail.com [reviewer]

Other contributors:

KangMin Ngo ngokangmin@gmail.com [reviewer]

Haley Overstreet OverstreetH@si.edu [reviewer]

CTFS-ForestGEO ForestGEO@si.edu [copyright holder]