Structure habitat data from measures of topography.

This function constructs habitat data based on elevation data. It calculates habitats in two steps: 1. It calculates mean elevation, convexity and slope for each quadrat (via [fgeo_topography()])). 2. It calculates habitats based on hierarchical clustering of the topographic metrics from step 1 (via [stats::hclust()]).

fgeo_habitat(elevation, gridsize, n, ...)

Arguments

elevation
One of these: * A list with at least three elements: `col` containing elevation data; and `xdim` and `ydim` giving plot dimensions; OR * A dataframe containing elevation data, in which case the parameters `xdim` and `ydim` must be provided. It may be the element `col` of a ForestGEO elevation-list or an object of class fgeo_elevation (see [fgeo.tool::fgeo_elevation()]).

gridsize
Number giving the size of each quadrat for which a habitat is calculated. Commonly, `gridsize = 20`.

n
Integer. Number of cluster-groups to construct (passed to the argument `k` to [stats::cutree()]).

...
Arguments passed to [fgeo_topography()].

Value

A dataframe of subclass fgeo_habitat, with columns `gx` and `gy`, rounded with accuracy determined by `gridsize`, and column `habitats`, with as many distinct integer values as determined by the argument `n`.

Input

The main input can be either the elevation list that ForestGEO delivers, or the element col of such list -- which is a dataframe containing the elevation data. Notice that the required arguments vary according to the main input (the elevation list or the elevation dataframe). Whatever the input, the dataframe containing the elevation data must have columns gx and gy or x and y.

See also

Examples

# Input a ForestGEO-like elevation list
elev_list <- fgeo.x::elevation
habs <- fgeo_habitat(elev_list, n = 4, gridsize = 20)
str(habs)

#> Classes 'fgeo_habitat', 'tbl_df', 'tbl' and 'data.frame':	400 obs. of  3 variables:
#>  $ gx      : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gy      : num  0 20 40 60 80 100 120 140 160 180 ...
#>  $ habitats: int  1 1 1 1 1 1 1 1 1 1 ...

# NOT RUN {
installed <- requireNamespace("fgeo.map", quietly = TRUE)
if (installed) {
  library(fgeo.map)
  autoplot(habs)
}

# A good use of habitat data is for calculating species-habitat associations
installed <- requireNamespace("fgeo.analyze", quietly = TRUE)
if (installed) {
  library(fgeo.analyze)

  elev_list <- fgeo.x::elevation
  habitat <- fgeo_habitat(elev_list, gridsize = 20, n = 4)
  census <- fgeo.analyze::luquillo_top3_sp
  species <- unique(census$sp)
  to_df(tt_test(census, species, habitat))
}

# If working with elevation dataframe you must provide xdim and ydim
elev_df <- fgeo.x::elevation$col
hab2 <- fgeo_habitat(elev_df, gridsize = 20, n = 4, xdim = 320, ydim = 500)
str(hab2)
# }