Torus Translation Test to determine habitat associations of tree species.

Determine habitat-species associations with code developed by Sabrina Russo, Daniel Zuleta, Matteo Detto, and Kyle Harms.

tt_test(census, sp, habitat, plotdim = NULL, gridsize = NULL)

Arguments

census
A dataframe; a ForestGEO _tree_ table (see details).

sp
Character sting giving any number of species-names.

habitat
Object giving the habitat designation for each plot partition defined by `gridsize`. See [`fgeo_habitat()`].

plotdim, gridsize
Plot dimensions and gridsize. If `NULL` (default) they will be guessed, and a message will inform you of the chosen values. If the guess is wrong, you should provide the correct values manually (and check that your habitat data is correct).

Value

A list of matrices. You can summarize the output with [summary()] and convert it to a dataframe with [to_df()]. You can also view the result with `View(your-result)`, and reduce it from a list of matrices to a single matix with `Reduce(rbind, your-result)`. See examples.

Details

You should only try to determine the habitat association for sufficiently abundant species. In a 50-ha plot, a minimum abundance of 50 trees/species has been used. Also, you should use data of individual trees (i.e. a tree table, and not a stem table with potentially multiple stems per tree. This test only makes sense at the population level. We are interested in knowing whether or not individuals of a species are aggregated on a habitat. Multiple stems of an individual do not represent population level processes but individual level processes.

Acknowledgments

Nestor Engone Obiang, David Kenfack, Jennifer Baltzer, and Rutuja Chitra-Tarak provided feedback. Daniel Zuleta provided guidance.

Interpretation of Output

* `N.Hab.1`: Count of stems of the focal species in habitat 1. * `Gr.Hab.1`: Count of instances the observed relative density of the focal species on habitat 1 was greater than the relative density based on the TT habitat map. * `Ls.Hab.1`: Count of instances the observed relative density of the focal species on habitat 1 was less than the relative density based on the TT habitat map. * `Eq.Hab.1`: Count of instances the observed relative density of the focal species on habitat 1 was equal to the relative density based on the TT habitat map. The sum of the `Gr.Hab.x`, `Ls.Hab.x`, and `Eq.Hab.x` columns for one habitat equals the number of 20 x20 quads in the plot. The `Rep.Agg.Neut` columns for each habitat indicate whether the species is significantly repelled (-1), aggregated (1), or neutrally distributed (0) on the habitat in question.

The probabilities associated with the test for whether these patterns are statistically significant are in the `Obs.Quantile` columns for each habitat. Note that to calculate the probability for repelled, it is the value given, but to calculate the probability for aggregated, it is 1 - the value given.

Values of the `Obs.Quantile` < 0.025 means that the species is repelled from that habitat, while values of the `Obs.Quantile` > 0.975 means that the species is aggregated on that habitat.

References

Zuleta, D., Russo, S.E., Barona, A. et al. Plant Soil (2018). https://doi.org/10.1007/s11104-018-3878-0.

See also

Examples

library(dplyr)

# Example data
tree <- luquillo_top3_sp
elevation <- fgeo.x::elevation

# Pick alive trees, of 10 mm or more
census <- filter(tree, status == "A", dbh >= 10)

# Pick sufficiently abundant species
pick <- filter(add_count(census, sp), n > 50)
species <- unique(pick$sp)

# Use your habitat data or create it from elevation data
habitat <- fgeo_habitat(elevation, gridsize = 20, n = 4)

# A list or matrices
tt_lst <- tt_test(census, species, habitat)

#> Using `plotdim = c(320, 500)`. To change this value see `?tt_test()`.

#> Using `gridsize = 20`. To change this value see `?tt_test()`.

tt_lst

#> [[1]]
#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> CASARB      29     1242      356        2              0        0.77625      20
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> CASARB      390     1206        4              0        0.24375      12
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> CASARB      778      817        5              0        0.48625       5
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> CASARB      932      658       10              0         0.5825
#> 
#> [[2]]
#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> PREMON      91     1093      504        3              0       0.683125      89
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> PREMON     1254      344        2              0        0.78375      40
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> PREMON      305     1292        3              0       0.190625      14
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> PREMON      270     1322        8              0        0.16875
#> 
#> [[3]]
#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> SLOBER      18      273     1324        3              0       0.170625      24
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> SLOBER      810      788        2              0        0.50625      17
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> SLOBER     1155      440        5              0       0.721875       7
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> SLOBER     1292      303        5              0         0.8075
#> 

# A simple summary to help you interpret the results
summary(tt_lst)

#>   Species Habitat_1 Habitat_2 Habitat_3 Habitat_4
#> 1  CASARB   neutral   neutral   neutral   neutral
#> 2  PREMON   neutral   neutral   neutral   neutral
#> 3  SLOBER   neutral   neutral   neutral   neutral

# A combined matrix
Reduce(rbind, tt_lst)

#>        N.Hab.1 Gr.Hab.1 Ls.Hab.1 Eq.Hab.1 Rep.Agg.Neut.1 Obs.Quantile.1 N.Hab.2
#> CASARB      29     1242      356        2              0       0.776250      20
#> PREMON      91     1093      504        3              0       0.683125      89
#> SLOBER      18      273     1324        3              0       0.170625      24
#>        Gr.Hab.2 Ls.Hab.2 Eq.Hab.2 Rep.Agg.Neut.2 Obs.Quantile.2 N.Hab.3
#> CASARB      390     1206        4              0        0.24375      12
#> PREMON     1254      344        2              0        0.78375      40
#> SLOBER      810      788        2              0        0.50625      17
#>        Gr.Hab.3 Ls.Hab.3 Eq.Hab.3 Rep.Agg.Neut.3 Obs.Quantile.3 N.Hab.4
#> CASARB      778      817        5              0       0.486250       5
#> PREMON      305     1292        3              0       0.190625      14
#> SLOBER     1155      440        5              0       0.721875       7
#>        Gr.Hab.4 Ls.Hab.4 Eq.Hab.4 Rep.Agg.Neut.4 Obs.Quantile.4
#> CASARB      932      658       10              0        0.58250
#> PREMON      270     1322        8              0        0.16875
#> SLOBER     1292      303        5              0        0.80750

# A dataframe
to_df(tt_lst)

#> # A tibble: 12 x 8
#>    habitat sp     N.Hab Gr.Hab Ls.Hab Eq.Hab Rep.Agg.Neut Obs.Quantile
#>  * <chr>   <chr>  <dbl>  <dbl>  <dbl>  <dbl>        <dbl>        <dbl>
#>  1 1       CASARB    29   1242    356      2            0        0.776
#>  2 2       CASARB    20    390   1206      4            0        0.244
#>  3 3       CASARB    12    778    817      5            0        0.486
#>  4 4       CASARB     5    932    658     10            0        0.582
#>  5 1       PREMON    91   1093    504      3            0        0.683
#>  6 2       PREMON    89   1254    344      2            0        0.784
#>  7 3       PREMON    40    305   1292      3            0        0.191
#>  8 4       PREMON    14    270   1322      8            0        0.169
#>  9 1       SLOBER    18    273   1324      3            0        0.171
#> 10 2       SLOBER    24    810    788      2            0        0.506
#> 11 3       SLOBER    17   1155    440      5            0        0.722
#> 12 4       SLOBER     7   1292    303      5            0        0.808